Don't worry about it. Read description if worrying.
Bascially, I took all our current opmodes... and moved them into a
folder. I just wanted to break them up so that they're out of the way
for new stuff -- plus I think that we should keep our code somewhat
organized. Secondly, I created a folder and opmode for testing something
that I'm working on. Thirdly, ask me if you want to know about a
template thing I did in Android Studio. I won't explain it here, but
it's going to be something that I'm using a lot, and it might help you
if you're interested.